+++ app/Http/Controllers/ProductController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
+use App\Models\Market;
 use App\Models\Product;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 
@@ -16,9 +18,10 @@ class ProductController extends Controller
 
     public function index()
     {
-        $this->authorize('admin_user');
+        $this->authorize('admin_has_market');
+        $market = User::find(auth()->user()->id);
         $data = [
-            'products' => Product::paginate(10),
+            'products' => Product::where('market_id', $market->market->id)->paginate(10),
         ];
         return view('admin.pages.products.index', $data);
     }
@@ -28,8 +31,12 @@ class ProductController extends Controller
      */
     public function create()
     {
-        $this->authorize('admin_user');
-        $data = [];
+        $this->authorize('admin_has_market');
+        $data = [
+            'categorys' => [
+                'Baju', 'Makanan', 'Minuman', 'Souvenir'
+            ]
+        ];
 
         return view('admin.pages.products.create', $data);
     }
@@ -39,15 +46,20 @@ class ProductController extends Controller
      */
     public function store(StoreProductRequest $request)
     {
-        $this->authorize('admin_user');
+        $this->authorize('admin_has_market');
+
+        // $market = Market::where('user_id', auth()->user()->id)->first();
+        $market = User::find(auth()->user()->id);
+        $market->load('market');
 
         $product = new Product();
-        $product->name = $request->name;
-        $product->price = $request->price;
-        $product->description = $request->description;
-        $product->stock = $request->stock;
-        $product->remainder = $request->remainder;
-
+        $product->name = $request->input('name');
+        $product->price = $request->input('price');
+        $product->category = $request->input('category');
+        $product->description = $request->input('description');
+        $product->stock = $request->input('stock');
+        $product->remainder = $request->input('stock');
+        $product->market_id = $market->market->id;
         if ($request->hasFile('image')) {
             $image_product = "product_" . time() . "." . $request->image->extension();
             $request->file('image')->move(public_path('image_products/'), $image_product);
@@ -65,7 +77,7 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
-        $this->authorize('admin_user');
+        $this->authorize('admin_has_market');
         $data = [
             'product' => $product,
         ];
@@ -78,9 +90,12 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
-        $this->authorize('admin_user');
+        $this->authorize('admin_has_market');
         $data = [
             'product' => $product,
+            'categorys' => [
+                'Baju', 'Makanan', 'Minuman', 'Souvenir'
+            ]
         ];
 
         return view('admin.pages.products.edit', $data);
@@ -91,20 +106,20 @@ class ProductController extends Controller
      */
     public function update(UpdateProductRequest $request, Product $product)
     {
-        $this->authorize('admin_user');
+        $this->authorize('admin_has_market');
 
-        $product->name = $request->name;
-        $product->price = $request->price;
-        $product->description = $request->description;
-        $product->stock = $request->stock;
-        $product->remainder = $request->remainder;
+        $product->name = $request->input('name');
+        $product->category = $request->input('category');
+        $product->price = $request->input('price');
+        $product->description = $request->input('description');
         if ($request->hasFile('image_new')) {
             if ($product->image != 'not_found') {
                 File::delete(public_path('image_products/' . $product->image));
+            } else {
+                $image_product = "product" . time() . "." . $request->image_new->extension();
+                $request->file('image_new')->move(public_path('image_products/'), $image_product);
+                $product->image = $image_product;
             }
-            $image_product = "product" . time() . "." . $request->image_new->extension();
-            $request->file('image_new')->move(public_path('image_products/'), $image_product);
-            $product->image = $image_product;
         }
         $product->save();
 
@@ -116,10 +131,22 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        $this->authorize('admin_has_market');
         if ($product->image !== "not_found") {
             unlink(public_path('image_products/' . $product->image));
         }
         $product->delete();
         return redirect('products')->with('success', 'Product deleted successfully');
+    }
+
+    public function destroyImage(Product $product)
+    {
+        $this->authorize('admin_has_market');
+        if ($product->image !== "not_found") {
+            unlink(public_path('image_products/' . $product->image));
+        }
+        $product->image = "not_found";
+        $product->save();
+        return redirect('products')->with('success', 'Product image deleted successfully');
     }
 }
