+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\UpdateUserRequest;
+use App\Models\Market;
 use App\Models\User;
 use Illuminate\Http\Request;
 
@@ -15,7 +16,7 @@ class UserController extends Controller
     {
         $this->authorize('superadmin_admin');
         $data = [
-            'users' => User::paginate(10),
+            'users' => User::whereNot('role', '1')->paginate(10),
         ];
         return view('admin.pages.users.index', $data);
     }
@@ -58,6 +59,10 @@ class UserController extends Controller
         $user = User::findOrFail($id);
         $data = [
             'user' => $user,
+            'roles' => [
+                '2' => 'Admin',
+                '3' => 'User'
+            ]
         ];
         return view('admin.pages.users.edit', $data);
     }
@@ -67,10 +72,18 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+
+        if (Market::where('user_id', $user->id)->exists()) {
+            if ($request->input('role') == '3') {
+                return redirect('/users')->with('error', 'User sudah terdaftar di sebagai admin toko!');
+            }
+        }
+
         $user->name = $request->input('name');
         $user->username = $request->input('username');
         $user->phone = $request->input('phone');
         $user->address = $request->input('address');
+        $user->role = $request->input('role');
         $user->save();
 
         return redirect('/users')->with('success', 'User berhasil diupdate!');
@@ -82,7 +95,11 @@ class UserController extends Controller
     public function destroy(string $id)
     {
         $this->authorize('superadmin');
+
         $user = User::findOrFail($id);
+        if (Market::where('user_id', $user->id)->exists()) {
+            return redirect('/users')->with('error', 'User sudah terdaftar di sebagai admin toko!');
+        }
         $user->delete();
         return redirect('/users')->with('error', 'User berhasil dihapus!');
     }
